.claude/plans/merge-convert-skills/tools/ir-synthesize-golang/pyproject.toml
[project]
name = "ir-synthesize-golang"
version = "0.1.0"
description = "Go synthesizer for the IR extraction/synthesis pipeline"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "AI Context Library Contributors" }
]
keywords = ["ir", "synthesis", "golang", "go", "code-generation"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Compilers",
    "Topic :: Software Development :: Code Generators",
    "Typing :: Typed",
]

dependencies = [
    "ir-core>=0.1.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.8.0",
    "ruff>=0.2.0",
    "black>=24.0.0",
]
all = [
    "ir-synthesize-golang[dev]",
]

[project.scripts]
ir-synthesize-golang = "ir_synthesize_golang.__main__:main"

[project.entry-points."ir_core.synthesizers"]
golang = "ir_synthesize_golang:GolangSynthesizer"

[project.urls]
Homepage = "https://github.com/your-org/ai-context-library"
Documentation = "https://your-org.github.io/ai-context-library/ir-synthesize-golang/"
Repository = "https://github.com/your-org/ai-context-library"
Issues = "https://github.com/your-org/ai-context-library/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ir_synthesize_golang"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
]

[tool.coverage.run]
source = ["ir_synthesize_golang"]
branch = true
omit = [
    "*/tests/*",
    "*/__main__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
    "if __name__ == .__main__.:",
]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_ignores = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
show_error_codes = true
show_column_numbers = true

[tool.ruff]
target-version = "py311"
line-length = 88
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # Pyflakes
    "I",    # isort
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "UP",   # pyupgrade
    "ARG",  # flake8-unused-arguments
    "SIM",  # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.isort]
known-first-party = ["ir_synthesize_golang", "ir_core"]

[tool.black]
target-version = ["py311"]
line-length = 88
